+++ includes/class-hrs-theme-setup.php
@@ -55,11 +55,13 @@ class HRS_Theme_Setup {
 	private function setup_hooks() {
 		add_action( 'after_setup_theme', array( $this, 'add_theme_support' ) );
 		add_action( 'after_setup_theme', array( $this, 'register_nav_menus' ) );
+		add_action( 'init', array( $this, 'register_taxonomies' ), 0 );
 
 		// Set Spine options.
 		add_action( 'after_setup_theme', array( $this, 'get_hrs_spine_schema' ), 5 );
 		add_filter( 'spine_get_title', array( $this, 'hrs_get_page_title' ) );
 		add_filter( 'spine_enable_builder_module', '__return_true' );
+		add_filter( 'spine_option_defaults', array( $this, 'hrs_spine_option_defaults' ) );
 	}
 
 	/**
@@ -74,6 +76,10 @@ class HRS_Theme_Setup {
 		require __DIR__ . '/shortcode-document-gallery.php';
 		// The HRS last updated label shortcode.
 		require __DIR__ . '/shortcode-last-updated.php';
+		// The HRS template tags.
+		require __DIR__ . '/hrs-template-tags.php';
+		// WP database queries.
+		require __DIR__ . '/hrs-queries.php';
 	}
 
 	/**
@@ -113,6 +119,100 @@ class HRS_Theme_Setup {
 	}
 
 	/**
+	 * Updates the Spine options list with HRS's default options.
+	 *
+	 * Overrides some of the default Spine options (typically set in the
+	 * Customizer) to align with the HRS child theme. The full HRS defaults are:
+	 *
+	 *   'spine_version'             => '2',
+	 *   'grid_style'                => 'fluid',
+	 *   'campus_location'           => '',
+	 *   'spine_color'               => 'white',
+	 *   'large_format'              => ' folio max-1386', // Max Width 1386px
+	 *   'theme_style'               => 'skeletal',
+	 *   'secondary_colors'          => 'default', // Crimson
+	 *   'theme_spacing'             => 'default', // 2em
+	 *   'global_main_header_sup'    => '',
+	 *   'global_main_header_sub'    => '',
+	 *   'main_header_show'          => true,
+	 *   'articletitle_show'         => true,
+	 *   'articletitle_header'       => false,
+	 *   'broken_binding'            => false,
+	 *   'bleed'                     => true,
+	 *   'search_state'              => 'closed',
+	 *   'crop'                      => true, // Cropped Spine (homepage)
+	 *   'spineless'                 => true, // Spineless (homepage) checked
+	 *   'open_sans'                 => '0', // Off
+	 *   'contact_name'              => 'Washington State University',
+	 *   'contact_department'        => 'Human Resource Services',
+	 *   'contact_url'               => 'https://hrs.wsu.edu/',
+	 *   'contact_streetAddress'     => 'PO Box 641014',
+	 *   'contact_addressLocality'   => 'Pullman, WA',
+	 *   'contact_postalCode'        => '99164',
+	 *   'contact_telephone'         => '509-335-4521',
+	 *   'contact_email'             => 'hrs@wsu.edu',
+	 *   'contact_ContactPoint'      => 'http://hrs.wsu.edu/hrs-contacts/',
+	 *   'contact_ContactPointTitle' => 'Contact Page...',
+	 *   'archive_content_display'   => 'full',
+	 *   'social_spot_one_type'      => 'facebook',
+	 *   'social_spot_one'           => 'https://www.facebook.com/wsuhrs',
+	 *   'social_spot_two_type'      => 'twitter',
+	 *   'social_spot_two'           => 'https://twitter.com/wsupullman',
+	 *   'social_spot_three_type'    => 'linkedin',
+	 *   'social_spot_three'         => 'https://www.linkedin.com/company/washington-state-university',
+	 *   'social_spot_four_type'     => 'directory',
+	 *   'social_spot_four'          => 'https://socialmedia.wsu.edu/',
+	 *   'post_social_placement'     => 'none',
+	 *   'show_author_page'          => '0',
+	 *   'show_breadcrumbs'          => 'top', // Only valid with Breadcrumb NavXT plugin installed.
+	 *   'front_page_title'          => false,
+	 *   'page_for_posts_title'      => false,
+	 *
+	 * @since 0.14.0
+	 *
+	 * @param array $spine_options The list of default Spine options.
+	 * @return array The updated list of default options.
+	 */
+	public function hrs_spine_option_defaults( $spine_options ) {
+
+		// Defaults for the spine options to merge with HRS defaults.
+		$spine_defaults = spine_get_option_defaults();
+
+		$hrs_defaults = array(
+			'grid_style'              => 'fluid',
+			'spine_color'             => 'white',
+			'theme_style'             => 'skeletal',
+			'large_format'            => ' folio max-1386',
+			'crop'                    => true,
+			'spineless'               => true,
+			'bleed'                   => true,
+			'open_sans'               => '0',
+			'contact_name'            => 'Washington State University',
+			'contact_department'      => 'Human Resource Services',
+			'contact_url'             => 'https://hrs.wsu.edu/',
+			'contact_streetAddress'   => 'PO Box 641014',
+			'contact_addressLocality' => 'Pullman, WA',
+			'contact_postalCode'      => '99164',
+			'contact_telephone'       => '509-335-4521',
+			'contact_email'           => 'hrs@wsu.edu',
+			'contact_ContactPoint'    => 'http://hrs.wsu.edu/hrs-contacts/',
+			'show_author_page'        => '0',
+			'social_spot_one_type'    => 'facebook',
+			'social_spot_one'         => 'https://www.facebook.com/wsuhrs',
+			'social_spot_two_type'    => 'twitter',
+			'social_spot_two'         => 'https://twitter.com/wsupullman',
+			'social_spot_three_type'  => 'linkedin',
+			'social_spot_three'       => 'https://www.linkedin.com/company/washington-state-university',
+			'social_spot_four_type'   => 'directory',
+			'social_spot_four'        => 'https://socialmedia.wsu.edu/',
+		);
+
+		$spine_options = wp_parse_args( $hrs_defaults, $spine_defaults );
+
+		return $spine_options;
+	}
+
+	/**
 	 * Adds theme support for features provided by WordPress.
 	 *
 	 * Gallery and caption HTML5 support is already added in the Spine parent
@@ -140,6 +240,45 @@ class HRS_Theme_Setup {
 			'hrs-search-menu' => __( 'Search Menu', 'hrs-wsu-edu' ),
 			'hrs-site-footer' => __( 'Site Footer', 'hrs-wsu-edu' ),
 		) );
+	}
+
+	/**
+	 * Creates the HRS taxonomies.
+	 *
+	 * Uses the WP taxonomy API to create custom taxonomy for the HRS site,
+	 * {@see register_taxonomy}.
+	 *
+	 * @since 0.14.0
+	 */
+	public function register_taxonomies() {
+
+		// Create the HRS Unit taxonomy.
+		$labels = array(
+			'name'              => _x( 'HRS Units', 'taxonomy general name', 'hrs-wsu-edu' ),
+			'singular_name'     => _x( 'HRS Unit', 'taxonomy singular name', 'hrs-wsu-edu' ),
+			'all_items'         => __( 'All Units', 'hrs-wsu-edu' ),
+			'edit_item'         => __( 'Edit Unit', 'hrs-wsu-edu' ),
+			'view_item'         => __( 'View Unit', 'hrs-wsu-edu' ),
+			'update_item'       => __( 'Update Unit', 'hrs-wsu-edu' ),
+			'add_new_item'      => __( 'Add New Unit', 'hrs-wsu-edu' ),
+			'new_item_name'     => __( 'New Unit Name', 'hrs-wsu-edu' ),
+			'parent_item'       => __( 'Parent Unit', 'hrs-wsu-edu' ),
+			'parent_item_colon' => __( 'HRS Unit: ', 'hrs-wsu-edu' ),
+			'search_items'      => __( 'Search Units', 'hrs-wsu-edu' ),
+		);
+
+		$args = array(
+			'hierarchical'      => true,
+			'labels'            => $labels,
+			'show_ui'           => true,
+			'show_admin_column' => true,
+			'query_var'         => true,
+			'rewrite'           => array(
+				'slug' => 'hrs-units',
+			),
+		);
+
+		register_taxonomy( 'hrs_unit', array( 'post', 'document' ), $args );
 	}
 
 	/**