+++ single.php
@@ -1,0 +1,36 @@
+<?php
+/**
+ * Single Post Template
+ *
+ * The template for displaying individual post views.
+ *
+ * @package WSU_Human_Resources_Services
+ * @since 0.14.0
+ */
+?>
+
+<?php get_header(); ?>
+
+<main id="wsuwp-main">
+
+	<?php
+
+	get_template_part( 'parts/headers' );
+
+	if ( function_exists( 'wsuwp_uc_get_object_type_slugs' ) && in_array( get_post_type(), wsuwp_uc_get_object_type_slugs(), true ) ) {
+		if ( 'wsuwp_uc_person' === get_post_type() ) {
+			get_template_part( 'parts/single-layout', 'wsuwp_uc_person' );
+		} else {
+			get_template_part( 'parts/single-layout', 'university-center' );
+		}
+	} else {
+		get_template_part( 'parts/single-layout', get_post_type() );
+	}
+
+	get_template_part( 'parts/footers' );
+
+	?>
+
+</main><!--/#page-->
+
+<?php get_footer();